Hide away the size of light values and metadata - These don't need to be bytes except when dealing with the implementation

+++ src/ddb/io/voxelnet/block/BlockAir.java
@@ -32,7 +32,7 @@ public class BlockAir extends Block
 	}
 	
 	@Override
-	public byte getOpacity()
+	public int getOpacity()
 	{
 		return 0;
 	}
